backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
psycopg2-binary
cryptography
fpdf2
pydantic
python-multipart
python-dotenv
